added getters/setters for lights in C++ and Lua

+++ ST_engine/src/main/game_manager/level/text.cpp
@@ -18,7 +18,7 @@
  * @param font The font to use (Must be already loaded).
  * @param font_size The size of the font.
  */
-ST::text::text(uint64_t id, int x, int y, SDL_Color color, const std::string& text_string, const std::string& font, int font_size) {
+ST::text::text(uint64_t id, int x, int y, SDL_Color color, const std::string& text_string, const std::string& font, uint8_t font_size) {
     this->ID = id;
     this->x = x;
     this->y = y;
